Add ladder capture and escape features, tests for same.
Ladder features are computed by is_ladder_capture() and
is_ladder_escape() in GameState.

A ladder capture is defined as a move by one player (the "hunter" in
the code) that leads to a forced capture of an adjacent enemy group
(the "prey") with two liberties before the move.

A ladder escape is defined as a move (by the prey) that brings a group
with one liberty to 3 or more liberties, or to 2 liberties neither of
which is a ladder_capture for the hunter against that prey group.

The two functions call each other recursively.  The prey group can be
specified in the call to each function, but the default is to treat
any adjacent group as potential prey.

Several tests, adapted from Fuego's code (LGPL license) are included,
along with a helper function in tests/parser_board.py to simplify
specifying tests via ascii boards.